+++ modules/android.nix
@@ -1,0 +1,11 @@
+{
+  nixosModule = { user, ... }: {
+    programs.adb.enable = true;
+
+    users.users.${user}.extraGroups = [ "adbusers" ];
+  };
+
+  homeModule = { pkgs, ... }: {
+    home.packages = builtins.attrValues { inherit (pkgs) scrcpy; };
+  };
+}