+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Added `android` module for using `adb` and `scrcpy`
 - Changed `nixos` to be `true` by default for Linux hosts
 - Refactored `if x then y else { }` into `optionalAttrs x y`
 - Refactored `hasSuffix "{darwin,linux}"` into `pkgs.stdenv.hostPlatform.is{Darwin,Linux}`